+++ src/routes/share.$gistId.tsx
@@ -176,7 +176,7 @@ function RouteComponent() {
                       navigate({
                         to: "/notes/$",
                         params: { _splat: userNote.id },
-                        search: { mode: "write", query: undefined },
+                        search: { query: undefined },
                       })
                     }}
                   >